Feature/blog

+++ Models/Product.cs
@@ -5,9 +5,9 @@
 		public int Id { get; set; }
 		public string Name { get; set; }
 		public string Description { get; set; }
-		public float Price { get; set; }
-		public float SalePrice { get; set; }
-		public byte Status { get; set; }
+		public double Price { get; set; }
+		public double SalePrice { get; set; }
+		public int Status { get; set; }
 		public string Image { get; set; }
 	}
 }
